Add PSA_CRYPTO_PLATFORM_ZEROIZE config option
Add a specific configuration option to provide
an alternate implementation of psa_crypto_platform_zeroize().

Dissociate it from PSA_CRYPTO_STD_FUNCTIONS. In an embedded
environment it is quite common to disable PSA_CRYPTO_STD_FUNCTIONS
and we do not want this to imply to provide an implementation of
psa_crypto_platform_zeroize() and such implementation is very
tricky.

Signed-off-by: Ronald Cron <ronald.cron@arm.com>