Modify how the configuration translation in extended
In PSA-Crypto, the configuration of the Mbed TLS
driver builtin implementation is deduced from the
PSA-Crypto configuration. The deduction logic
is an extension of what is done by
config_adjust_legacy_from_psa.h in Mbed TLS.
Modify how the extension is done such as it
does not rely on an additional commit on top of
Mbed TLS development.

Signed-off-by: Ronald Cron <ronald.cron@arm.com>