Have a low-level error code override any high-level one
Let the low-level error code win because that is usually the right choice,
e.g. to propagate out-of-memory errors through semantic errors, or to
propagate an ASN.1 parsing error through certificate verification.

We mostly use MBEDTLS_ERROR_ADD with both a low-level and a high-level
error, but we also occasionally use it with a low-level code that can be 0.
Ensure that the value is nonzero whenever either value is nonzero.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>