Store the algorithm for a PK context in the context
Attach an algorithm or algorithm policy (leaving the hash part unspecified)
to every PK context. By default, it is the same algorithm that
`mbedtls_pk_sign()` and `mbedtls_pk_verify()` were already using, with one
exception: for contexts of type `MBEDTLS_PK_RSA`, the sign/verify functions
used to honor the choice of padding mode in the underlying
`mbedtls_rsa_context` object, but now default to PKCS#1v1.5. To use PSS,
call `mbedtls_pk_set_algorithm()` instead (or in addition) to setting the
padding mode in the RSA context.

The new function `mbedtls_pk_set_algorithm()` can override the default
choice. If so, `mbedtls_pk_get_psa_attributes()` uses the algorithm set in
the context, if it is compatible with the requested usage.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>