config: remove definition of MBEDTLS_NO_PLATFORM_ENTROPY
MBEDTLS_PLATFORM_GET_ENTROPY_ALT is meant to be an alternative to
platform entropy support, not an additional option as it was in the
past. Therefore enabling MBEDTLS_PLATFORM_GET_ENTROPY_ALT has the
same effect of enabling MBEDTLS_NO_PLATFORM_ENTROPY, so the latter can
be removed.

This commit only removes the definition of MBEDTLS_NO_PLATFORM_ENTROPY
from the configuration files. Library code and test adaptations will
come in follow-up commits.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>