cmake: Move link of Windows platform libraries to core
The original driver for this move is that the CMake
versions on the CI do not support to link libraries
to an object library.

Beyond that, the Windows libraries are platform
libraries not libraries needed by builtin drivers
directly. Thus linking them directly to the
tfpsacrypto library does also make sense.

Signed-off-by: Ronald Cron <ronald.cron@arm.com>