Allow NULL pointers for zero-length output buffers
A (NULL, 0) output buffer is represented as the pair (*p = NULL, start =
NULL) and is a valid thing to pass as an output buffer to represent a
zero-length output.

We should always return MBEDTLS_ERR_ASN1_BUF_TOO_SMALL in this case as
the minimum length of an encoded integer is > 0 bytes.

This allows us to test a zero-length output buffer as well.

Signed-off-by: David Horstmann <david.horstmann@arm.com>