Fix bug when dealing with the value zero
Leading-zero checking caused us to record a zero-length integer when
dealing with the value zero, which then led to a one-byte buffer
overread when checking if the value had a leading '1' bit.

Since we know how to encode zero, treat it as a special case and encode
it as one byte.

Signed-off-by: David Horstmann <david.horstmann@arm.com>