Rename the changed mbedtls_pk_can_do_ext to mbedtls_pk_can_do_psa
The function isn't used much by third parties. Give the new function a
different name: this will ease the migration in TLS code that uses it in a
way whose semantics is changing.

Give the new flag for derive-public a temporary name and value, and link to
the relevant PSA API discussion.

Note that since we aren't preserving the old function, this is not critical
for the 1.0 release.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>