tests: pk: add tests for mbedtls_pk_can_do_psa()
Adding tests for the new "mbedtls_pk_can_do_psa()" function. Basic testing
is added to the usual "test_suite_pk.data" file, while extended testing
is added to a new "test_suite_pk.pk_can_do_psa" file. Reason for a
separate file there is that a lot of test cases are being added and it's
clearer to keep them separate from the default PK testing data file.

"test_suite_pk.pk_can_do_psa" was generated by means of a python script
which is also included in this commit.
I didn't copy/pasted the script in the commit message as done for other
similar cases in the past because also the script is quite long and I
thought it's easier for reviewers to have it in clear to check it more
easily.

Note: currently there's also a test_number parameter that's added to
the generated test cases. This is only meant to be used to help debugging
in order to stop to a specific test case without need to comment out
other ones or other tricks. This can be removed later.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>