Change mutex methods from function pointers to functions
Changing the mutex methods at runtime is not done in normal operation.
Remove this possibility. Now the mutex methods must be provided as direct
functions. This removes a bit of complexity inside the library and removes
some of the startup-time burden in applications.

We do change the mutex methods in our test instrumentation, to keep detect
problematic sequences of calls. A subsequent commit will make this possible
through test hooks.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>