scripts: driver_templates: fix few driver wrappers
Several driver wrapper functions don't generate the right dispatch code
for opaque drivers. Including the "OS-template-opaque" template fixes
this, as it generates code that calls the corresponding opaque driver
by matching it with the key's location.

The functions fixed in this commit seem to be the necessary minimum for
an opaque driver to implement TLS1.3 support.

Signed-off-by: Ilie Halip <ilie.halip@nxp.com>