psa: add support for PSA_KEY_USAGE_DERIVE_PUBLIC
PSA_KEY_USAGE_DERIVE_PUBLIC is meant to be added to public PSA APIs in
the future, but for the time being the new macro definition is added
to "crypto_extra.h".
Crypto core is also updated to allow the creation of a PSA key which
supports this new usage flag.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>