Refactor MD HMAC initialization in the library
Change internal uses of mbedtls_md_setup() that use hmac=1 to instead
pass hmac=0 as required and do the initialization of HMAC in a separate
call to mbedtls_md_hmac_setup().

Note that since we have separated the two allocations of memory, if the
initialization of HMAC fails, we can no longer clean up the rest of the
allocated context. As a result, in some usages an extra call to
mbedtls_md_free() is needed to clean up the non-HMAC state if something
goes wrong in HMAC allocation.

Signed-off-by: David Horstmann <david.horstmann@arm.com>