Change ECP trace counters from self tests to test hooks
In `ecp.c`, the generic arithmetic code can optionally keep track of how
many calls it makes to certain auxiliary functions (scalar-scalar
multiplication in `mbedtls_mpi_mul_mod()`, point addition in
`ecp_add_mixed()` and doubling in `ecp_double_jac()`). The counters are kept
in a global variable.

Until now this was done whenever `MBEDTLS_SELF_TEST` was enabled, and only
checked by the self-test code. The usage of the counters was not
thread-safe, which not only meant that self-tests could not be run in
parallel with any ECC operation, but also that TSan would complain whenever
ECC arithmetic was done in a build with `MBEDTLS_SELF_TEST` enabled (because
the counters would get corrupted, although we only cared about those
counters in the self tests).

Change this to use the mechanism for test hooks with `MBEDTLS_TEST_HOOKS`.
Now the counter updates are only done if `MBEDTLS_TEST_HOOKS` is enabled and
some non-user-serviceable pointer is non-null, following the general
principles for `MBEDTLS_TEST_HOOKS`. Also, following these general
principles, we only set the hook in unit test code, not when running the
self tests in production.

Fixes Mbed-TLS/TF-PSA-Crypto#443.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>