Free the context if mbedtls_md_hmac_setup() fails
On a failure of mbedtls_md_hmac_setup(), free the parent
mbedtls_md_context, even though we did not allocate it. This is okay
because mbedtls_md_hmac_setup() is exclusively used internally and
always directly after calling mbedtls_md_setup(). If HMAC setup fails,
we count that as a failure of the whole setup of MD.

Signed-off-by: David Horstmann <david.horstmann@arm.com>