Rectify tuning of the configuration for TSan
`MBEDTLS_ECP_RESTARTABLE` does not need to be disabled. Restartable ECC is
not thread-safe, but this does not affect the thread safety of the ECC code
as long as you don't try restartable operations (including interruptible PSA
functions).

The reason why `MBEDTLS_SELF_TEST` needs to be disabled is not that the
self-tests are not thread-safe, but that enabling them causes the ECC code
to not be thread-safe. The race condition only affects internal trace
counters that only the self-tests care about, but that is sufficient to get
a deserved complaint from TSan.
https://github.com/Mbed-TLS/TF-PSA-Crypto/issues/443

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>