Do not include context types in client-only mode
When building with MBEDTLS_PSA_CRYPTO_CLIENT but not
MBEDTLS_PSA_CRYPTO_C, do not include the definitions of context types
specific to the builtin implementation.

These are not needed or used, as the build is 'client-only', meaning
that operation structs contain an mbedtls_psa_client_handle_t rather
than the real operation state.

Signed-off-by: David Horstmann <david.horstmann@arm.com>