Use -fzero-init-padding-bits=all to avoid GCC 15.0 "feature"
GCC 15 interprets the C23 union initialization rules to allow it to
skip initializing portions of structs and unions, this breaks
code like:

	const struct psa_pake_operation_s v = { 0 };

as large portions of 'v' will be left uninitialized.

To avoid this "feature", use -fzero-init-padding-bits=all to restore
GCC behavior to that in earlier versions of the compiler.

Here's a similar change as adopted in Linux:

https://lore.kernel.org/lkml/20250127191031.245214-3-kees@kernel.org/

Signed-off-by: Keith Packard <keithp@keithp.com>