Add MBEDTLS_EMPTY_INITIALIZER macro for GCC 15 and clang 18 fixes
C23 defines the empty initializer, {}, with different semantics than a
partial initializer containing only zero values (like {0}):

 "If an object that has automatic storage duration is initialized with
  an empty initializer, its value is the same as the initialization of
  a static storage duration object." §6.7.10

Reading further, we find that this will effectively set the entire
object to zero, including padding.

GCC 15 (and some builds of clang 18) start leaving portions of structs
uninitialized unless we use empty initializers. Fortunately they
both support empty initializers even when not compiling in c23 mode.

Because empty initializers are a new feature in C23, we cannot use
them with previous versions of language, so this new macro takes a
parameter to use as the initializer value for the first composite
member when building with a pre-C23 compiler.

One unfortunate effect of using this macro is that we cannot use
explicit initializer values for any members lest we lose the empty
initializer semantics for the whole object. This places an implicit
dependency on the representation for any symbolic
initializers. Capturing that in some fashion (perhaps using a static
assertion) might be a nice enhancement to this change.

Signed-off-by: Keith Packard <keithp@keithp.com>