Support legacy encoding of PSA_ALG_JPAKE in key policies
This is for the sake of persistent keys created with Mbed TLS 3.x.

In this commit, the library still allows `PSA_ALG_JPAKE_BETA` in the
policies of newly created keys. This greatly simplifies the library code and
the tests. However, in the documentation, don't promise that this will
remain the case.

Fixes https://github.com/Mbed-TLS/TF-PSA-Crypto/issues/505

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>