Exercise JPAKE only if specified in the test data
The automatic test case generator generates some JPAKE tests. They must not
be exercised, since the test case generator is not aware of the limitations
on the hash algorithms. Honor `TEST_FLAG_EXERCISE`, which is not set on the
problematic test cases. Do set `TEST_FLAG_EXERCISE` on the manually written
test cases that we do want to exercise.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>