asymmetric_signature_key_policy: pass validity through usage flags
In `asymmetric_signature_key_policy` tests, express whether a given usage is
valid through a mask of usage flags. Formerly this was expressed as a mix of
test data (`payload_length_arg=0` to convey that no usage was allowed) and
logic in the code (to determine which usage was allowed).

Now we always use a valid payload length, even in tests that expect a policy
violation.

No change in test coverage of the library.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>