Test policy with ECDSA variants
Deterministic ECDSA with a hash, randomized ECDSA with a hash, and the
special case `PSA_ALG_DETERMINISTIC_ECDSA`, are three separate families of
algorithms, and the policy determines which one to use. Test that this is
enforced.

As a special case, `PSA_ALG_ECDSA(PSA_ALG_ANY_HASH)` allows
`PSA_ALG_DETERMINISTIC_ECDSA`. This is our historical behavior, but no
longer allowed by the PSA specification.
https://github.com/Mbed-TLS/TF-PSA-Crypto/issues/533

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>