Use instruction-level parallelism with AES-NI
One feature of the AES-NI instructions is instruction-level parallelism,
which means that, while one round of AES takes multiple clock cycles to
complete, another instruction can already be executed on the next clock
cycle. Only when the result of the AES round is used do we need to wait
for it to finish.

In particular, this makes it possible to encrypt or decrypt multiple AES
blocks in parallel which can be used to speed up AES in some modes of
operation. In this commit, it is used to speed up AES-CTR.

Signed-off-by: Max Fillinger <maximilian.fillinger@sentyron.com>