Fix bug with mbedtls_aesce_crypt_ecb
The function mbedtls_aes_crypt_ecb_multiblock() didn't exit after
it was done encrypting the input with mbedtls_aesce_crypt_ecb() but
instead went on to encrypt the input buffer again with the AES software
implementation. This did not only waste time, but also created incorrect
outputs when trying to encrypt a buffer in place.

Signed-off-by: Max Fillinger <maximilian.fillinger@sentyron.com>