Avoid inlining mbedtls_internal_aes_encrypt()
The tests in "mbedtls/tests/scripts/all.sh test_aesni" check that the
software implementation of AES is compiled by checking whether this
function name appears in aes.o. This function is static, and my changes
in the previous commits cause it to be inlined by GCC 5.4.0, which is
the compiler used in the test image in mbedtls-test. This does not seem
to be the case with more recent versions.

To make the tests pass, I've added the noinline attribute to the
function.

Signed-off-by: Max Fillinger <maximilian.fillinger@sentyron.com>